# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ContaFácil API — gestão contábil
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clientes:
    get:
      operationId: listarClientes
      tags: [clientes]
      summary: Lista todos os clientes
      responses:
        "200":
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cliente"
    post:
      operationId: criarCliente
      tags: [clientes]
      summary: Cria ou atualiza um cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClienteInput"
      responses:
        "200":
          description: Cliente salvo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
  /clientes/{id}:
    patch:
      operationId: atualizarCliente
      tags: [clientes]
      summary: Atualiza campos avulsos de um cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientePatch"
      responses:
        "200":
          description: Cliente atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
    delete:
      operationId: removerCliente
      tags: [clientes]
      summary: Apaga um cliente sem histórico (só admin). Com competência gerada, use inativar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "204":
          description: Removido
        "409":
          $ref: "#/components/responses/Erro"
  /clientes/{id}/inativar:
    post:
      operationId: inativarCliente
      tags: [clientes]
      summary: Inativa o cliente (sai das competências novas; o histórico fica)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Cliente inativado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
  /clientes/{id}/reativar:
    post:
      operationId: reativarCliente
      tags: [clientes]
      summary: Reativa um cliente inativado
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Cliente reativado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
  /clientes/{id}/segredos:
    get:
      operationId: getSegredosCliente
      tags: [clientes]
      summary: Revela as senhas gov.br e NFS-e do cliente (fica registrado na auditoria)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Senhas em claro
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SegredosCliente"
        "403":
          $ref: "#/components/responses/Erro"
        "404":
          $ref: "#/components/responses/Erro"
...
components:
  securitySchemes:
    tokenJobs:
      type: http
      scheme: bearer
      description: "`Authorization: Bearer ${TOKEN_JOBS}`"
  responses:
    Erro:
      description: Erro tratado pela API
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Erro"
  schemas:
...
    Erro:
      type: object
      required: [error, codigo]
      properties:
        error: { type: string }
        codigo: { type: string }
        details: {}
        issues:
          type: array
          items: { type: object, additionalProperties: true }
...
    SegredosCliente:
      type: object
      required: [senhaGov, senhaNfse]
      properties:
        senhaGov: { type: ["string", "null"] }
        senhaNfse: { type: ["string", "null"] }
...
    Cliente:
      type: object
      required: [id, razaoSocial, ativo, obrigacoes, temSenhaGov, temSenhaNfse]
      properties:
        id: { type: integer }
        codigo: { type: ["integer", "null"], minimum: 0 }
        razaoSocial: { type: string, minLength: 1, maxLength: 200 }
        cnpj: { type: ["string", "null"] }
        cnaePrincipal: { type: ["string", "null"] }
        regime: { $ref: "#/components/schemas/RegimeTributario" }
        inscricaoEstadual: { type: ["string", "null"] }
        inscricaoMunicipal: { type: ["string", "null"] }
        formaEnvio: { type: ["string", "null"] }
        procuracao: { type: ["string", "null"] }
        procuracaoVencimento: { type: ["string", "null"], format: date }
        socioNome: { type: ["string", "null"] }
        socioCpf: { type: ["string", "null"] }
        # As senhas nunca saem na listagem: só por /clientes/{id}/segredos, com auditoria.
        temSenhaGov: { type: boolean }
        temSenhaNfse: { type: boolean }
        observacao: { type: ["string", "null"] }
        valorHonorario: { type: ["string", "null"], pattern: '^-?\d+(\.\d{1,2})?$' }
        diaVencimentoHonorario: { type: ["integer", "null"], minimum: 1, maximum: 31 }
        contatoNome: { type: ["string", "null"] }
        whatsapp: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        ativo: { type: boolean }
        inativadoEm: { type: ["string", "null"], format: date-time }
        obrigacoes:
          type: array
          items: { type: integer }
    ClienteInput:
      type: object
      required: [razaoSocial]
      properties:
        id: { type: ["integer", "null"] }
        codigo: { type: ["integer", "null"], minimum: 0 }
        razaoSocial: { type: string, minLength: 1, maxLength: 200 }
        cnpj: { type: ["string", "null"] }
        cnaePrincipal: { type: ["string", "null"] }
        regime: { $ref: "#/components/schemas/RegimeTributario" }
        inscricaoEstadual: { type: ["string", "null"] }
        inscricaoMunicipal: { type: ["string", "null"] }
        formaEnvio: { type: ["string", "null"] }
        procuracao: { type: ["string", "null"] }
        procuracaoVencimento: { type: ["string", "null"], format: date }
        socioNome: { type: ["string", "null"] }
        socioCpf: { type: ["string", "null"] }
        senhaGov: { type: ["string", "null"] }
        senhaNfse: { type: ["string", "null"] }
        observacao: { type: ["string", "null"] }
        valorHonorario: { type: ["string", "null"], pattern: '^-?\d+(\.\d{1,2})?$' }
        diaVencimentoHonorario: { type: ["integer", "null"], minimum: 1, maximum: 31 }
        contatoNome: { type: ["string", "null"] }
        whatsapp: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        ativo: { type: boolean }
        obrigacoes:
          type: array
          items: { type: integer }
    RegimeTributario:
      type: ["string", "null"]
      enum: [simples_nacional, mei, lucro_presumido, lucro_real, null]

    # Edição célula-a-célula na tela de Dados cadastrais: todos os campos são
    # opcionais e só o que vier no body é gravado (não mexe nas obrigações).
    ClientePatch:
      type: object
      properties:
        codigo: { type: ["integer", "null"], minimum: 0 }
        razaoSocial: { type: string, minLength: 1, maxLength: 200 }
        cnpj: { type: ["string", "null"] }
        cnaePrincipal: { type: ["string", "null"] }
        regime: { $ref: "#/components/schemas/RegimeTributario" }
        inscricaoEstadual: { type: ["string", "null"] }
        inscricaoMunicipal: { type: ["string", "null"] }
        formaEnvio: { type: ["string", "null"] }
        procuracao: { type: ["string", "null"] }
        procuracaoVencimento: { type: ["string", "null"], format: date }
        socioNome: { type: ["string", "null"] }
        socioCpf: { type: ["string", "null"] }
        senhaGov: { type: ["string", "null"] }
        senhaNfse: { type: ["string", "null"] }
        observacao: { type: ["string", "null"] }
        valorHonorario: { type: ["string", "null"], pattern: '^-?\d+(\.\d{1,2})?$' }
        diaVencimentoHonorario: { type: ["integer", "null"], minimum: 1, maximum: 31 }
        contatoNome: { type: ["string", "null"] }
        whatsapp: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        ativo: { type: boolean }
...